requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
web3==6.11.0
ipfshttpclient==0.8.0
python-dotenv==1.0.0
Werkzeug==2.3.7
pytest==7.4.2
black==23.9.1
flake8==6.1.0
gunicorn==21.2.0